management-api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
email-validator
python-multipart
stripe
sendgrid
prometheus_client